pyproject.toml
[project]
name = "invoice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.2",
    "docx2pdf>=0.1.8",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "python-docx>=1.2.0",
]
